refactor(test-support): InMemoryFlamingockBulder and InMemoryAuditStore javdocs
